% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_log_odba.R
\name{acf_plot_log_odba}
\alias{acf_plot_log_odba}
\title{Create ACF plots from log ODBA data}
\usage{
acf_plot_log_odba(data, filename)
}
\arguments{
\item{data}{A dataframe with column ODBA}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Creates image file containing an ACF plot for log ODBA.
}
