% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_static.R
\name{acf_plot_static}
\alias{acf_plot_static}
\title{Create ACF plots from static data}
\usage{
acf_plot_static(data, filename)
}
\arguments{
\item{data}{A dataframe with columns X_static,
Y_static, Z_static}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
Image file containing ACF plots for X_static, Y_static, Z_static
}
\description{
Create ACF plots from static data
}
