% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{ar_densities_labelled}
\alias{ar_densities_labelled}
\title{Returns densities for autoregressive model, for some states known}
\usage{
ar_densities_labelled(x, mod, m, q, n, state)
}
\arguments{
\item{x}{Vector of observations}

\item{mod}{List of parameters}

\item{m}{Number of states}

\item{n}{Number of observations}

\item{state}{List of state values, if provided. 0 represents an unknown state value.}
}
\value{
n x m matrix of state dependent probability densities
}
\description{
Returns densities for autoregressive model, for some states known
}
