% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{ar_hmm_generate_sample}
\alias{ar_hmm_generate_sample}
\title{Generate sample from HMM with autoregressive model}
\usage{
ar_hmm_generate_sample(ns, mod)
}
\arguments{
\item{ns}{Sample size}

\item{mod}{List of HMM parameters}
}
\value{
Dataframe including index, state, obs
}
\description{
Generate sample from HMM with autoregressive model
}
