% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{ar_hmm_lbackward}
\alias{ar_hmm_lbackward}
\title{Get backward probabilities}
\usage{
ar_hmm_lbackward(x, mod)
}
\arguments{
\item{x}{Vector of observations}

\item{mod}{List of HMM parameters}
}
\value{
Matrix of backward probabilities
}
\description{
Get backward probabilities
}
