% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{ar_hmm_lforward}
\alias{ar_hmm_lforward}
\title{Get forward probabilities}
\usage{
ar_hmm_lforward(x, mod)
}
\arguments{
\item{x}{Vector of observations}

\item{mod}{List of HMM parameters}
}
\value{
Matrix of forward probabilities
}
\description{
Get forward probabilities
}
