% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{ar_hmm_mle}
\alias{ar_hmm_mle}
\title{Maximum likelihood estimation of univariate autoregresive parameters}
\usage{
ar_hmm_mle(
  x,
  m,
  q,
  mu0,
  sigma0,
  gamma0,
  phi0,
  delta0 = NULL,
  stationary = TRUE,
  hessian = FALSE
)
}
\arguments{
\item{x}{Vector of observations}

\item{m}{Number of states}

\item{mu0}{Vector of length m, initial values for means the white noise}

\item{sigma0}{Vector of length m, initial values for standard deviations}

\item{gamma0}{Matrix of size m x m, initial values for transition probability matrix}

\item{phi0}{Matrix of size m x q, initial values for autoregressive parameters}

\item{delta0}{Optional, vector of length m, initial values for
initial distribution}

\item{stationary}{Boolean, whether the HMM is stationary or not}

\item{hessian}{Boolean, whether to return the inverse hessian}
}
\value{
List of results
}
\description{
Maximum likelihood estimation of univariate autoregresive parameters
}
