% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{ar_hmm_pseudo_residuals}
\alias{ar_hmm_pseudo_residuals}
\title{Generate pseudo residuals}
\usage{
ar_hmm_pseudo_residuals(x, mod, type, stationary)
}
\arguments{
\item{x}{Vector of observations}

\item{mod}{List of HMM parameters}

\item{type}{Type of pseudo-residual, either "ordinary" or "forecast"}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
Dataframe of pseudo-residuals, observations, index
}
\description{
Generate pseudo residuals
}
