% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{ar_hmm_pw2pn}
\alias{ar_hmm_pw2pn}
\title{Transform normal working parameters to natural parameters}
\usage{
ar_hmm_pw2pn(m, q, parvect, stationary = TRUE)
}
\arguments{
\item{m}{Number of states}

\item{q}{Order of the autoregressive model}

\item{parvect}{Vector of working parameters}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
List of natural parameters mu, sigma, gamma, phi, delta
}
\description{
Transform normal working parameters to natural parameters
}
