% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{ar_hmm_viterbi}
\alias{ar_hmm_viterbi}
\title{Global decoding of states}
\usage{
ar_hmm_viterbi(x, mod)
}
\arguments{
\item{x}{Vector of observations}

\item{mod}{List of HMM parameters}
}
\value{
Dataframe of decoded states and index
}
\description{
Global decoding of states
}
