% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{ar_jacobian}
\alias{ar_jacobian}
\title{Get Jacobian matrix}
\usage{
ar_jacobian(mod, n)
}
\arguments{
\item{mod}{List of maximum likelihood estimation results}

\item{n}{Total number of working parameters (excluding delta)}
}
\value{
Jacobian matrix, size n x n
}
\description{
Get Jacobian matrix
}
