% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_log_odba.R
\name{behavior_hist_log_odba}
\alias{behavior_hist_log_odba}
\title{Create histograms from log ODBA data, for each behavior interval}
\usage{
behavior_hist_log_odba(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Behavior, ODBA}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Each image file contains histograms aggregating log ODBA
over each behavior interval for a given behavior.
A behavior interval is a time interval in which the behavior remains
constant.
Each behavior interval is labelled by a number, in chronological order.
If there are many behavior intervals for a given behavior, the
plot may be hard to read.
}
