% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_dynamic.R
\name{behavior_pairs_plot_dynamic}
\alias{behavior_pairs_plot_dynamic}
\title{Create correlation plots of X_dynamic, Y_dynamic, and Z_dynamic,
divided by behavior}
\usage{
behavior_pairs_plot_dynamic(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Behavior, X_dynamic,
Y_dynamic, Z_dynamic, ODBA}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
Several image plots of correlation plots, divided by behavior
}
\description{
Create correlation plots of X_dynamic, Y_dynamic, and Z_dynamic,
divided by behavior
}
\examples{
filename <- "Custom_Lady_27Mar17_dynamic.csv"
data <- read.csv(filename)
behavior_pairs_plot_dynamic(data, "Lady_27Mar17_dynamic")
}
