% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_static.R
\name{behavior_pairs_plot_static}
\alias{behavior_pairs_plot_static}
\title{Create correlation plots of X_static, Y_static, and Z_static, divided
by behavior}
\usage{
behavior_pairs_plot_static(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Behavior, X_static,
Y_static, Z_static}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
Several image plots of correlation plots, divided by behavior
}
\description{
Create correlation plots of X_static, Y_static, and Z_static, divided
by behavior
}
