% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_log_odba.R
\name{behavior_plot_log_odba}
\alias{behavior_plot_log_odba}
\title{Create line plots from log ODBA data, which are colored by behavior}
\usage{
behavior_plot_log_odba(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Time, Behavior, ODBA}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Creates image file containing a line plot for log ODBA, colored
by behavior.
}
\examples{
filename <- "Custom_Lady_27Mar17_dynamic.csv"
data <- read.csv(filename)
labelled_data <- data \%>\% filter(!is.na(Behavior))
behavior_plot_log_odba(labelled_data, "Lady_27Mar17_plot_behavior")
}
