% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_static.R
\name{behavior_plot_static}
\alias{behavior_plot_static}
\title{Create line plots from static data, which are colored by behavior}
\usage{
behavior_plot_static(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Time, Behavior, X_static,
Y_static, Z_static}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
An image file containing line plots for X_static, Y_static, Z_static.
Each line plot is colored by behavior.
}
\description{
Create line plots from static data, which are colored by behavior
}
