% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{diag_exp}
\alias{diag_exp}
\title{Exponentiate diagonal elements}
\usage{
diag_exp(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
Matrix with diagonal elements exponentiated.
}
\description{
Exponentiate diagonal elements
}
\examples{
diag_exp(diag(2, 2))
}
