% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{diag_log_lower}
\alias{diag_log_lower}
\title{Get log diagonal and lower triangular entries}
\usage{
diag_log_lower(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
Vector of log diagonal elements
and lower triangular entries of the matrix,
going down column by column.
}
\description{
Applies log to only diagonal elements of matrix.
}
\examples{
diag_log_lower(matrix(c(1:4), nrow=2))

}
