% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_dynamic.R
\name{filtered_hist_dynamic}
\alias{filtered_hist_dynamic}
\title{Create histograms from dynamic and ODBA data, aggregated by behavior}
\usage{
filtered_hist_dynamic(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Behavior, X_dynamic,
Y_dynamic, Z_dynamic, ODBA}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Creates several image files containing histograms for
X_dynamic, Y_dynamic, Z_dynamic, and ODBA aggregated by behavior.
One set of images includes histograms for all behaviors in one
plot, divided by data type. Another set includes X_dynamic, Y_dynamic,
and Z_dynamic histograms in one image and ODBA histogram in another,
divided by behavior.
}
\examples{
filename <- "Custom_Lady_27Mar17_dynamic.csv"
data <- read.csv(filename)
labelled_data <- data \%>\% filter(!is.na(Behavior))
filtered_hist_dynamic(labelled_data, "Lady_27Mar17_histogram_filtered")
}
