% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_log_odba.R
\name{filtered_hist_log_odba}
\alias{filtered_hist_log_odba}
\title{Create histograms from log ODBA data, aggregated by behavior}
\usage{
filtered_hist_log_odba(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Behavior, ODBA}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Creates several image files containing histograms for
log ODBA, aggregated by behavior. One image includes histograms
for all behaviors in one plot. The remaining images contain a histogram
for log ODBA for a single behavior.
}
\examples{
filename <- "Custom_Lady_27Mar17_dynamic.csv"
data <- read.csv(filename)
labelled_data <- data \%>\% filter(!is.na(Behavior))
filtered_hist_log_odba(labelled_data, "Lady_27Mar17_histogram_filtered")
}
