% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_static.R
\name{filtered_hist_static}
\alias{filtered_hist_static}
\title{Create histograms from static data, aggregated by behavior}
\usage{
filtered_hist_static(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Behavior, X_static,
Y_static, Z_static}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Creates several image files containing histograms for
X_static, Y_static, Z_static aggregated by behavior.
One set of images includes histograms for all behaviors in one
plot, divided by data type. Another set includes X_static,
Y_static, Z_static histograms in one image, divided by behavior.
}
\examples{
filename <- "Custom_Lady_27Mar17_static.csv"
data <- read.csv(filename)
labelled_data <- data \%>\% filter(!is.na(Behavior))
filtered_hist_static(labelled_data, "Lady_27Mar17_histogram_filtered")
}
