% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{get_all_ar_means}
\alias{get_all_ar_means}
\title{Get all means for autoregressive series}
\usage{
get_all_ar_means(mu, phi, m, q, x)
}
\arguments{
\item{mu}{Vector of length m, containing means for the white noise in each
state dependent distribution}

\item{phi}{Matrix of size m x q, containing autoregressive parameters. Each row contains
the parameters corresponding to a single state.}

\item{m}{Number of states}

\item{q}{Order of the autoregressive model}

\item{x}{Vector of observations coming from an autoregressive series}
}
\value{
n x m matrix (where n is the number of observations)
containing means for the autoregressive model
}
\description{
Get all means for autoregressive series
}
