% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/indep_multivariate_autoregressive_hmm_functions.R
\name{get_all_inmar_means}
\alias{get_all_inmar_means}
\title{Get all means for multivariate autoregressive series}
\usage{
get_all_inmar_means(x, mod, m, q, k)
}
\arguments{
\item{x}{Observations coming from an independent multivariate autoregressive series,
in a matrix with k rows. Each row corresponds to a variable.}

\item{mod}{List of HMM parameters}

\item{m}{Number of states}

\item{q}{Order of the autoregressive model}

\item{k}{Number of variables}
}
\value{
List of autoregressive means corresponding to
each state. Means are in a n x k matrix, where
n is the number of observations.
}
\description{
Get all means for multivariate autoregressive series
}
