% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_sd.R
\name{get_all_sd_data}
\alias{get_all_sd_data}
\title{Create multiple files containing dynamic and static standard deviation data}
\usage{
get_all_sd_data(names)
}
\arguments{
\item{names}{List of names of data sets}
}
\value{
None
}
\description{
Create multiple files containing dynamic and static standard deviation data
}
