% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_autoregressive_hmm_functions.R
\name{get_ar_mean}
\alias{get_ar_mean}
\title{Get mean corresponding to a given index in an autoregressive series}
\usage{
get_ar_mean(mu, phi, x, q, i)
}
\arguments{
\item{mu}{Vector of length m, containing means for the white noise in each
state dependent distribution}

\item{phi}{Matrix of size m x q, containing autoregressive parameters. Each row contains
the parameters corresponding to a single state.}

\item{x}{Vector of observations coming from an autoregressive series}

\item{q}{Order of the autoregressive model}

\item{i}{Index of the desired mean}
}
\value{
Vector of length m containing means corresponding to index i
for the given autoregressive model
}
\description{
Get mean corresponding to a given index in an autoregressive series
}
