% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data.R
\name{get_custom_data}
\alias{get_custom_data}
\title{Combine dynamic or static data with behavior timestamps}
\usage{
get_custom_data(data_file, timestamps_file, filename)
}
\arguments{
\item{data_file}{Name of file containing dynamic or static data}

\item{timestamps_file}{Name of file containing behavior timestamps}

\item{filename}{Name of new file created}
}
\value{
Creates new csv file with combined data.
}
\description{
If date and time are in same column, the data is split into separate
Date and Time columns.
New columns are created for Behavior and Prey info.
}
\examples{
get_custom_data("Lady_27Mar17_Dynamic25Hz.csv",
"TimestampedData_Lady_27Mar17.csv",
"Custom_Lady_27Mar17_dynamic.csv")
}
