% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_sd.R
\name{get_dynamic_sd_data}
\alias{get_dynamic_sd_data}
\title{Get file of standard deviations of dynamic data}
\usage{
get_dynamic_sd_data(old_filename, filename)
}
\arguments{
\item{old_filename}{Name of csv file containing dynamic data. Must have
columns Time, X_dynamic, Y_dynamic, Z_dynamic, Behavior, Prey.}

\item{filename}{Name of new csv file containing standard deviations}
}
\value{
None
}
\description{
Standard deviations are taken every second (25 data points).
}
\examples{
old_filename <- "Custom_BigDaddy_3Apr17_dynamic.csv"
filename <- "Mean_BigDaddy_3Apr17_dynamic.csv"
get_dynamic_sd_data(old_filename, filename)
}
