% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_means.R
\name{get_log_odba_mean_plots}
\alias{get_log_odba_mean_plots}
\title{Create multiple plots for log ODBA mean data}
\usage{
get_log_odba_mean_plots(names)
}
\arguments{
\item{names}{List of strings, containing the names used to identify the
data sets}
}
\value{
None
}
\description{
Creates multiple image files, with different line plots, histograms, etc.
}
\examples{
names <- c("BigDaddy_3Apr17", "BigDaddy_20Mar17",
"BigGuy_15Feb18", "Eliza_7Sept17",
"Eliza_20Sept17", "Lady_27Mar17")
get_log_odba_mean_plots(names)
}
