% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_dynamic.R
\name{get_plots_dynamic}
\alias{get_plots_dynamic}
\title{Create multiple plots for dynamic and ODBA data}
\usage{
get_plots_dynamic(names)
}
\arguments{
\item{names}{List of strings, containing the names used to identify the
data sets}
}
\value{
Multiple image files
}
\description{
Create multiple plots for dynamic and ODBA data
}
