% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_log_odba.R
\name{get_plots_log_odba}
\alias{get_plots_log_odba}
\title{Create multiple plots for log ODBA data}
\usage{
get_plots_log_odba(names)
}
\arguments{
\item{names}{List of strings, containing the names used to identify the
data sets}
}
\value{
None
}
\description{
Creates multiple image files, with different line plots, histograms, etc.
}
