% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_static.R
\name{get_plots_static}
\alias{get_plots_static}
\title{Create multiple plots for static data}
\usage{
get_plots_static(names)
}
\arguments{
\item{names}{List of strings, containing the names used to identify the
data sets}
}
\value{
Multiple image files
}
\description{
Create multiple plots for static data
}
