% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_log_sd.R
\name{get_static_log_sd_data}
\alias{get_static_log_sd_data}
\title{Get file of log standard deviations of static data}
\usage{
get_static_log_sd_data(old_filename, filename)
}
\arguments{
\item{old_filename}{Name of csv file containing static data}

\item{filename}{Name of new csv file containing standard deviations}
}
\value{

}
\description{
Log standard deviations are taken every second (25 data points).
}
