% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_log_sd.R
\name{get_static_log_sd_plots}
\alias{get_static_log_sd_plots}
\title{Create multiple plots for static log standard deviation data}
\usage{
get_static_log_sd_plots(names)
}
\arguments{
\item{names}{List of strings, containing the names used to identify the
data sets}
}
\value{
None
}
\description{
Creates multiple image files, with different line plots, histograms, etc.
}
