% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_means.R
\name{get_static_mean_plots}
\alias{get_static_mean_plots}
\title{Create multiple plots for static mean data}
\usage{
get_static_mean_plots(names)
}
\arguments{
\item{names}{List of strings, containing the names used to identify the
data sets}
}
\value{
None
}
\description{
Creates multiple image files, with different line plots, histograms, etc.
}
