% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_hmm.R
\name{graph_hmm_hist}
\alias{graph_hmm_hist}
\title{Creates histogram from data generated from univariate HMM}
\usage{
graph_hmm_hist(output)
}
\arguments{
\item{output}{Dataframe with columns index, obs}
}
\value{
ggplot
}
\description{
Creates histogram from data generated from univariate HMM
}
