% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_dynamic.R
\name{hist_plot_dynamic}
\alias{hist_plot_dynamic}
\title{Create histograms from dynamic and ODBA data}
\usage{
hist_plot_dynamic(data, filename)
}
\arguments{
\item{data}{A dataframe with columns X_dynamic,
Y_dynamic, Z_dynamic, ODBA}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Creates one image file containing histograms for X_dynamic, Y_dynamic, Z_dynamic,
and another image file containing a histogram for ODBA.
}
\examples{
filename <- "Custom_Lady_27Mar17_dynamic.csv"
data <- read.csv(filename)
hist_plot_dynamic(data, "Lady_27Mar17_histogram")
}
