% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/indep_multivariate_autoregressive_hmm_functions.R
\name{inmar_bootstrap_estimates}
\alias{inmar_bootstrap_estimates}
\title{Get bootstrapped estimates of parameters}
\usage{
inmar_bootstrap_estimates(mod, n, len, stationary)
}
\arguments{
\item{mod}{List of maximum likelihood estimation results}

\item{n}{Number of bootstrap samples}

\item{len}{Number of observations}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
List of estimates
}
\description{
Get bootstrapped estimates of parameters
}
