% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/indep_multivariate_autoregressive_hmm_functions.R
\name{inmar_dist_mat}
\alias{inmar_dist_mat}
\title{Get multivariate autoregressive distribution function}
\usage{
inmar_dist_mat(x, mod, n)
}
\arguments{
\item{x}{Matrix of observations, rows represent each variable}

\item{mod}{List of maximum likelihood estimation results}

\item{n}{Number of observations}
}
\value{
Matrix of multivariate autoregressive probabilities
}
\description{
Get multivariate autoregressive distribution function
}
