% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/indep_multivariate_autoregressive_hmm_functions.R
\name{inmar_hmm_mllk}
\alias{inmar_hmm_mllk}
\title{Get negative log-likelihood from the working parameters}
\usage{
inmar_hmm_mllk(parvect, x, m, q, k, stationary = TRUE, state = NULL)
}
\arguments{
\item{parvect}{Vector of working parameters}

\item{x}{Matrix of observations, rows represent each variable}

\item{m}{Number of states}

\item{stationary}{Boolean, whether the HMM is stationary or not}

\item{state}{List of state values, if provided. 0 represents an unknown state value.}
}
\value{
Negative log-likelihood
}
\description{
Get negative log-likelihood from the working parameters
}
