% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/indep_multivariate_autoregressive_hmm_functions.R
\name{inmar_hmm_pn2pw}
\alias{inmar_hmm_pn2pw}
\title{Transform multivariate autoregressive natural parameters to working parameters}
\usage{
inmar_hmm_pn2pw(m, mu, sigma, gamma, phi, delta = NULL, stationary = TRUE)
}
\arguments{
\item{m}{Number of states}

\item{mu}{List of vectors of length m, means for white noise in each
state dependent distribution}

\item{sigma}{List of vectors of length m, standard deviations
for each state dependent multivariate normal distribution}

\item{gamma}{Transition probabiilty matrix, size m x m}

\item{phi}{List of k x q matrices, each row contains the autoregressive
parameters for a variable.}

\item{delta}{Optional, vector of length m containing
initial distribution}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
Vector of working parameters
}
\description{
mu and phi do not need to be transformed, as there are no constraints.
}
