% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_multivariate_normal_hmm_functions.R
\name{inmvnorm_densities}
\alias{inmvnorm_densities}
\title{Get matrix of state dependent probability densities}
\usage{
inmvnorm_densities(x, mod, m, k, n)
}
\arguments{
\item{x}{Vector containing one observation}

\item{mod}{List of parameters}

\item{m}{Number of states}

\item{k}{Number of variables}

\item{n}{Number of observations}
}
\value{
n x m matrix of state dependent probability densities
}
\description{
Get matrix of state dependent probability densities
}
