% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_multivariate_normal_hmm_functions.R
\name{inmvnorm_densities_labelled}
\alias{inmvnorm_densities_labelled}
\title{Get matrix of state dependent probability densities}
\usage{
inmvnorm_densities_labelled(x, mod, m, k, n, state)
}
\arguments{
\item{x}{Vector containing one observation}

\item{mod}{List of parameters}

\item{m}{Number of states}

\item{k}{Number of variables}

\item{n}{Number of observations}

\item{state}{List of state values, if provided. 0 represents an unknown state value.}
}
\value{
n x m matrix of state dependent probability densities
}
\description{
Get matrix of state dependent probability densities
}
