% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_multivariate_normal_hmm_functions.R
\name{inmvnorm_dist_mat}
\alias{inmvnorm_dist_mat}
\title{Get multivariate normal distribution function}
\usage{
inmvnorm_dist_mat(x, mod, n)
}
\arguments{
\item{x}{Matrix of observations, rows represent each variable}

\item{mod}{List of maximum likelihood estimation results}

\item{n}{Number of observations}
}
\value{
Matrix of multivariate normal probabilities
}
\description{
Get multivariate normal distribution function
}
