% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_multivariate_normal_hmm_functions.R
\name{inmvnorm_hmm_pn2pw}
\alias{inmvnorm_hmm_pn2pw}
\title{Transform independent multivariate normal natural parameters to working parameters}
\usage{
inmvnorm_hmm_pn2pw(m, mu, sigma, gamma, delta = NULL, stationary = TRUE)
}
\arguments{
\item{m}{Number of states}

\item{mu}{List of vectors of length m, means for each
state dependent multivariate normal distribution}

\item{sigma}{List of vectors of length m, standard deviations
for each state dependent multivariate normal distribution}

\item{gamma}{Transition probabiilty matrix, size m x m}

\item{delta}{Optional, vector of length m containing
initial distribution}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{

}
\description{
mu does not need to be transformed, as there are no constraints.
}
