% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_multivariate_normal_hmm_functions.R
\name{inmvnorm_hmm_sample_one}
\alias{inmvnorm_hmm_sample_one}
\title{Generate one sample from HMM with independent multivariate normal distribution}
\usage{
inmvnorm_hmm_sample_one(state, mod)
}
\arguments{
\item{state}{State the HMM is in}

\item{mod}{List of parameters}
}
\value{
Vector containing generated sample
}
\description{
Generate one sample from HMM with independent multivariate normal distribution
}
