% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indep_multivariate_normal_hmm_functions.R
\name{inmvnorm_jacobian}
\alias{inmvnorm_jacobian}
\title{Get Jacobian matrix}
\usage{
inmvnorm_jacobian(mod, n)
}
\arguments{
\item{mod}{List of maximum likelihood estimation results}

\item{n}{Total number of working parameters (excluding delta)}
}
\value{
Jacobian matrix
}
\description{
Get Jacobian matrix
}
