% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_log_odba.R
\name{line_plot_log_odba}
\alias{line_plot_log_odba}
\title{Create line plots from log ODBA data}
\usage{
line_plot_log_odba(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Time, ODBA}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Creates image file containing a line plot for log ODBA.
}
