% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_static.R
\name{line_plot_static}
\alias{line_plot_static}
\title{Create line plots from static data}
\usage{
line_plot_static(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Time, X_static,
Y_static, Z_static}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Creates an image file containing line plots
for X_static, Y_static, Z_static.
}
\examples{
filename <- "Custom_Lady_27Mar17_static.csv"
data <- read.csv(filename)
line_plot_static(data, "Lady_27Mar17_line_plot")
}
