% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_autoregressive_hmm_functions.R
\name{mar_bootstrap_ci}
\alias{mar_bootstrap_ci}
\title{Confidence intervals for estimated parameters by bootstrapping}
\usage{
mar_bootstrap_ci(mod, bootstrap, alpha)
}
\arguments{
\item{mod}{Maximum likelihood estimates of parameters}

\item{bootstrap}{Bootstrapped estimates for parameters}

\item{alpha}{Confidence level}
}
\value{
List of lower and upper bounds for confidence intervals
for each parameter
}
\description{
Confidence intervals for estimated parameters by bootstrapping
}
