% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_autoregressive_hmm_functions.R
\name{mar_densities}
\alias{mar_densities}
\title{Get matrix of state dependent probability densities}
\usage{
mar_densities(x, mod, m, q, k, n)
}
\arguments{
\item{x}{Matrix of observations, rows represent each variable}

\item{mod}{List of parameters}

\item{m}{Number of states}

\item{n}{Number of observations}
}
\value{
n x m matrix of state dependent probability densities
}
\description{
Get matrix of state dependent probability densities
}
