% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_autoregressive_hmm_functions.R
\name{mar_hmm_generate_sample}
\alias{mar_hmm_generate_sample}
\title{Generate samples from HMM with multivariate autoregressive model}
\usage{
mar_hmm_generate_sample(ns, mod)
}
\arguments{
\item{ns}{Number of samples}

\item{mod}{List of model parameters}
}
\value{
List including vector of indices, vector of states,
and k x ns matrix containing generated samples
(where k is the number of variables)
}
\description{
Generate samples from HMM with multivariate autoregressive model
}
