% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_autoregressive_hmm_functions.R
\name{mar_hmm_lforward}
\alias{mar_hmm_lforward}
\title{Get forward probabilities}
\usage{
mar_hmm_lforward(x, mod)
}
\arguments{
\item{x}{Matrix of observations, rows represent each variable}

\item{mod}{List of maximum likelihood estimation results}
}
\value{
Matrix of forward probabilities
}
\description{
Get forward probabilities
}
