% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_autoregressive_hmm_functions.R
\name{mar_hmm_pn2pw}
\alias{mar_hmm_pn2pw}
\title{Transform multivariate autoregressive natural parameters to working parameters}
\usage{
mar_hmm_pn2pw(m, mu, sigma, gamma, phi, delta = NULL, stationary = TRUE)
}
\arguments{
\item{m}{Number of states}

\item{mu}{List of vectors of length m, means for white noise in each
state dependent distribution}

\item{sigma}{List of matrices of size m x m, covariance matrices
for each state dependent distribution}

\item{gamma}{Transition probabiilty matrix, size m x m}

\item{phi}{List of k x (k x q) matrices, containing the autoregressive
parameters. Each matrix corresponds to a state. The first k x k entries
are the parameters for index i - 1, and so on up to index i - q.}

\item{delta}{Optional, vector of length m containing
initial distribution}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
Vector of working parameters
}
\description{
mu and phi do not need to be transformed, as there are no constraints.
We only need to transform diagonal elements of sigma, since there
are no constraints on the covariances.
Include only the lower triangular and diagional elements
of the sigma matrix, since covariance matrices must be symmetric.
}
