% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_autoregressive_hmm_functions.R
\name{mar_hmm_pw2pn}
\alias{mar_hmm_pw2pn}
\title{Transform multivariate autoregressive working parameters to natural parameters}
\usage{
mar_hmm_pw2pn(m, q, k, parvect, stationary = TRUE)
}
\arguments{
\item{m}{Number of states}

\item{parvect}{Vector of working parameters}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
List of natural parameters
}
\description{
Transform multivariate autoregressive working parameters to natural parameters
}
