% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_autoregressive_hmm_functions.R
\name{mar_hmm_viterbi}
\alias{mar_hmm_viterbi}
\title{Global decoding of states}
\usage{
mar_hmm_viterbi(x, mod)
}
\arguments{
\item{x}{Matrix of observations, rows represent each variable}

\item{mod}{List of maximum likelihood estimation results}
}
\value{
Dataframe of decoded states and index
}
\description{
Global decoding of states
}
