% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_autoregressive_hmm_functions.R
\name{mar_inv_hessian}
\alias{mar_inv_hessian}
\title{Get inverse of hessian matrix}
\usage{
mar_inv_hessian(mod, stationary = TRUE)
}
\arguments{
\item{mod}{List of maximum likelihood estimation results}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
Inverse hessian matrix
}
\description{
Transform hessian associated with working parameters
outputted by nlm.
If not stationary, exclude values associated with delta parameter
from the hessian matrix.
}
