% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_autoregressive_hmm_functions.R
\name{mar_jacobian}
\alias{mar_jacobian}
\title{Get Jacobian matrix}
\usage{
mar_jacobian(mod, n)
}
\arguments{
\item{mod}{List of maximum likelihood estimation results}

\item{n}{Total number of working parameters (excluding delta)}
}
\value{
Jacobian matrix
}
\description{
Get Jacobian matrix
}
