% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_bootstrap_ci}
\alias{mvnorm_bootstrap_ci}
\title{Confidence intervals for estimated parameters by bootstrapping}
\usage{
mvnorm_bootstrap_ci(mod, bootstrap, alpha, m, k)
}
\arguments{
\item{mod}{Maximum likelihood estimates of parameters}

\item{bootstrap}{Bootstrapped estimates for parameters}

\item{alpha}{Confidence level}

\item{m}{Number of states}

\item{k}{Number of variables}
}
\value{
List of lower and upper bounds for confidence intervals
for each parameter
}
\description{
Confidence intervals for estimated parameters by bootstrapping
}
