% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_densities}
\alias{mvnorm_densities}
\title{Get matrix of state dependent probability densities}
\usage{
mvnorm_densities(x, mod, m, n)
}
\arguments{
\item{x}{Vector containing one observation}

\item{mod}{List of parameters}

\item{m}{Number of states}

\item{n}{Number of observations}
}
\value{
n x m matrix of state dependent probability densities
}
\description{
Get matrix of state dependent probability densities
}
