% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_dist_mat}
\alias{mvnorm_dist_mat}
\title{Get multivariate normal distribution function}
\usage{
mvnorm_dist_mat(x, mod)
}
\arguments{
\item{x}{Matrix of observations, rows represent each variable}

\item{mod}{List of parameters}
}
\value{
Matrix of multivariate normal probabilities
}
\description{
Get multivariate normal distribution function
}
