% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_hmm_generate_sample}
\alias{mvnorm_hmm_generate_sample}
\title{Generate samples from HMM with multivariate normal distribution}
\usage{
mvnorm_hmm_generate_sample(ns, mod)
}
\arguments{
\item{ns}{Number of samples}

\item{mod}{List of model parameters}
}
\value{
k * ns matrix containing generated samples,
where k is the number of variables
}
\description{
Generate samples from HMM with multivariate normal distribution
}
