% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_hmm_lbackward}
\alias{mvnorm_hmm_lbackward}
\title{Get backward probabilities}
\usage{
mvnorm_hmm_lbackward(x, mod)
}
\arguments{
\item{x}{Matrix of observations, rows represent each variable}

\item{mod}{List of maximum likelihood estimation results}
}
\value{
Matrix of backward probabilities
}
\description{
Get backward probabilities
}
