% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_hmm_lforward}
\alias{mvnorm_hmm_lforward}
\title{Get forward probabilities}
\usage{
mvnorm_hmm_lforward(x, mod)
}
\arguments{
\item{x}{Matrix of observations, rows represent each variable}

\item{mod}{List of parameters}
}
\value{
Matrix of forward probabilities
}
\description{
Get forward probabilities
}
