% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_hmm_mle}
\alias{mvnorm_hmm_mle}
\title{Maximum likelihood estimation of normal parameters}
\usage{
mvnorm_hmm_mle(
  x,
  m,
  k,
  mu0,
  sigma0,
  gamma0,
  delta0 = NULL,
  stationary = TRUE,
  hessian = FALSE
)
}
\arguments{
\item{x}{Matrix of observations, rows represent each variable}

\item{m}{Number of states}

\item{k}{Number of variables}

\item{mu0}{List of vectors of length m, initial values for means}

\item{sigma0}{List of matrices of size m x m,
initial values for covariance matrices}

\item{gamma0}{Initial values for ransition probabiilty matrix, size m x m}

\item{delta0}{Optional, vector of length m containing initial values
initial distribution}

\item{stationary}{Boolean, whether the HMM is stationary or not}

\item{hessian}{Boolean, whether to return the inverse hessian}
}
\value{
List of results
}
\description{
Maximum likelihood estimation of normal parameters
}
