% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_hmm_mllk}
\alias{mvnorm_hmm_mllk}
\title{Get negative log-likelihood from the working parameters}
\usage{
mvnorm_hmm_mllk(parvect, x, m, k, stationary = TRUE)
}
\arguments{
\item{parvect}{Vector of working parameters}

\item{x}{Matrix of observations, rows represent each variable}

\item{m}{Number of states}

\item{k}{Number of variables}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
Negative log-likelihood
}
\description{
Get negative log-likelihood from the working parameters
}
