% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_hmm_pseudo_residuals}
\alias{mvnorm_hmm_pseudo_residuals}
\title{Generate pseudo residuals}
\usage{
mvnorm_hmm_pseudo_residuals(x, mod, type, stationary = TRUE)
}
\arguments{
\item{x}{Matrix of observations, rows represent each variable}

\item{mod}{List of maximum likelihood estimation results}

\item{type}{Type of pseudo-residual, either "ordinary" or "forecast"}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
Dataframe of pseudo-residuals, observations, index
}
\description{
Generate pseudo residuals
}
