% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_hmm_pw2pn}
\alias{mvnorm_hmm_pw2pn}
\title{Transform multivariate normal working parameters to natural parameters}
\usage{
mvnorm_hmm_pw2pn(m, k, parvect, stationary = TRUE)
}
\arguments{
\item{m}{Number of states}

\item{k}{Number of variables}

\item{parvect}{Vector of working parameters}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
List of natural parameters
}
\description{
Transform multivariate normal working parameters to natural parameters
}
