% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_hmm_functions.R
\name{mvnorm_hmm_sample_one}
\alias{mvnorm_hmm_sample_one}
\title{Generate one sample from HMM with multivariate normal distribution}
\usage{
mvnorm_hmm_sample_one(state, mod)
}
\arguments{
\item{state}{State the HMM is in}

\item{mod}{List of parameters}
}
\value{
Vector containing generated sample
}
\description{
Generate one sample from HMM with multivariate normal distribution
}
