% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_bootstrap_ci}
\alias{norm_bootstrap_ci}
\title{Confidence intervals for estimated parameters by bootstrapping}
\usage{
norm_bootstrap_ci(mod, bootstrap, alpha, m)
}
\arguments{
\item{mod}{Maximum likelihood estimates of parameters}

\item{bootstrap}{List of bootstrap estimates of parameters}

\item{alpha}{Confidence level}

\item{m}{Number of states}
}
\value{
List of lower and upper bounds for confidence intervals
for each parameter
}
\description{
Confidence intervals for estimated parameters by bootstrapping
}
