% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_bootstrap_covariance}
\alias{norm_bootstrap_covariance}
\title{Estimate covariance matrix of parameters from bootstrap estimates}
\usage{
norm_bootstrap_covariance(bootstrap, m, n)
}
\arguments{
\item{bootstrap}{List of bootstrap estimates of parameters}

\item{m}{Number of states}

\item{n}{Number of bootstrap samples}
}
\value{
Estimate of covariance matrix of parameters
}
\description{
Estimate covariance matrix of parameters from bootstrap estimates
}
