% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_densities}
\alias{norm_densities}
\title{Returns normal densities}
\usage{
norm_densities(x, mod, m, n)
}
\arguments{
\item{x}{Vector of observations}

\item{m}{Number of states}

\item{n}{Number of observations}
}
\value{
Matrix of normal densities, dimensions n x m
}
\description{
Returns normal densities
}
