% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_hmm_lforward}
\alias{norm_hmm_lforward}
\title{Get forward probabilities}
\usage{
norm_hmm_lforward(x, mod)
}
\arguments{
\item{x}{Vector of observations}

\item{mod}{List of maximum likelihood estimation results}
}
\value{
Matrix of forward probabilities
}
\description{
Get forward probabilities
}
