% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_hmm_mle}
\alias{norm_hmm_mle}
\title{Maximum likelihood estimation of normal parameters}
\usage{
norm_hmm_mle(
  x,
  m,
  mu0,
  sigma0,
  gamma0,
  delta0 = NULL,
  stationary = TRUE,
  hessian = FALSE
)
}
\arguments{
\item{x}{Vector of observations}

\item{m}{Number of states}

\item{mu0}{Vector of length m, initial value for means}

\item{sigma0}{Vector of length m, initial value for standard deviations}

\item{gamma0}{Matrix of size m x m, initial value for transition probability matrix}

\item{delta0}{Optional, vector of length m, initial value for
initial distribution}

\item{stationary}{Boolean, whether the HMM is stationary or not}

\item{hessian}{Boolean, whether to return the inverse hessian}
}
\value{
List of results
}
\description{
Maximum likelihood estimation of normal parameters
}
