% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_hmm_mllk}
\alias{norm_hmm_mllk}
\title{Get negative log-likelihood from the working parameters}
\usage{
norm_hmm_mllk(parvect, x, m, stationary = TRUE)
}
\arguments{
\item{x}{Vector of observations}

\item{m}{Number of states}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
Negative log-likelihood
}
\description{
Get negative log-likelihood from the working parameters
}
