% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_hmm_pn2pw}
\alias{norm_hmm_pn2pw}
\title{Transform normal natural parameters to working parameters}
\usage{
norm_hmm_pn2pw(m, mu, sigma, gamma, delta = NULL, stationary = TRUE)
}
\arguments{
\item{m}{Number of states}

\item{mu}{Vector of length m, containing means for each
state dependent normal distribution}

\item{sigma}{Vector of length m, containing standard
deviations for each state dependent normal distribution}

\item{gamma}{Transition probabiilty matrix, size m x m}

\item{delta}{Optional, vector of length m containing
initial distribution}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
Vector of working parameters
}
\description{
Transform normal natural parameters to working parameters
}
