% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_hmm_pw2pn}
\alias{norm_hmm_pw2pn}
\title{Transform normal working parameters to natural parameters}
\usage{
norm_hmm_pw2pn(m, parvect, stationary = TRUE)
}
\arguments{
\item{m}{Number of states}

\item{parvect}{Vector of working parameters}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
List of natural parameters mu, sigma, gamma, delta
}
\description{
Transform normal working parameters to natural parameters
}
