% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_hmm_viterbi}
\alias{norm_hmm_viterbi}
\title{Global decoding of states}
\usage{
norm_hmm_viterbi(x, mod)
}
\arguments{
\item{x}{Vector of observations}

\item{mod}{List of maximum likelihood estimation results}
}
\value{
Dataframe of decoded states and index
}
\description{
Global decoding of states
}
