% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_jacobian}
\alias{norm_jacobian}
\title{Get Jacobian matrix}
\usage{
norm_jacobian(m, n, parvect, stationary = TRUE)
}
\arguments{
\item{m}{Number of states}

\item{n}{Total number of working parameters (excluding delta)}

\item{parvect}{Vector of working parameters}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
Jacobian matrix
}
\description{
Get Jacobian matrix
}
