% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_normal_hmm_functions.R
\name{norm_marginal}
\alias{norm_marginal}
\title{Get univariate normal marginal distribution
from an HMM.}
\usage{
norm_marginal(start, end, n, mod, stationary)
}
\arguments{
\item{start}{Smallest value in range.}

\item{end}{Largest value in range.}

\item{mod}{List of HMM parameters}

\item{stationary}{Boolean, whether the HMM is stationary or not}
}
\value{
Dataframe
}
\description{
Get univariate normal marginal distribution
from an HMM.
}
