% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_log_odba.R
\name{pacf_plot_log_odba}
\alias{pacf_plot_log_odba}
\title{Create PACF plots from log ODBA data}
\usage{
pacf_plot_log_odba(data, filename)
}
\arguments{
\item{data}{A dataframe with column ODBA}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Creates image file containing a PACF plot for log ODBA.
}
\examples{
filename <- "Custom_Lady_27Mar17_dynamic.csv"
data <- read.csv(filename)
pacf_plot_log_odba(data, "Lady_27Mar17_pacf")
}
