% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_hmm.R
\name{pseudo_residual_plot}
\alias{pseudo_residual_plot}
\title{Pseudo residual plots}
\usage{
pseudo_residual_plot(data)
}
\arguments{
\item{data}{Dataframe including index, npsr (normal pseudo-residuals)}
}
\value{
ggplot, including index plot, histogram, qq plot, acf plot
}
\description{
Pseudo residual plots
}
