% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{triangular_num}
\alias{triangular_num}
\title{Get nth triangular number}
\usage{
triangular_num(n)
}
\arguments{
\item{n}{A non-negative integer.}
}
\value{
The n-th triangular number. (Where the 0th
triangular number is 0.)
}
\description{
Get nth triangular number
}
\examples{
triangular_num(3)
}
