% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_static.R
\name{behavior_hist_static}
\alias{behavior_hist_static}
\title{Create histograms from static data, for each behavior interval}
\usage{
behavior_hist_static(data, filename)
}
\arguments{
\item{data}{A dataframe with columns Behavior, X_static,
Y_static, Z_static}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
Several image files. Each image file contains histograms
aggregating a given data type (X_static, Y_static, Z_static)
over a given behavior.
}
\description{
A behavior interval is a time interval in which the behavior remains
constant.
Each behavior interval is labelled by a number, in chronological order.
Each plot contains histograms aggregating a given data type over
a given behavior.
If there are many behavior intervals for a given behavior, the
plot may be hard to read.
}
