% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_log_sd.R
\name{get_all_log_sd_data}
\alias{get_all_log_sd_data}
\title{Create multiple files containing dynamic and static log standard deviation data}
\usage{
get_all_log_sd_data(names)
}
\arguments{
\item{names}{List of names of data sets}
}
\value{
None
}
\description{
Create multiple files containing dynamic and static log standard deviation data
}
\examples{
names <- c("BigDaddy_3Apr17", "BigDaddy_20Mar17",
"BigGuy_15Feb18", "Eliza_7Sept17",
"Eliza_20Sept17", "Lady_27Mar17")
get_all_log_sd_data(names)
}
