% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_means.R
\name{get_static_mean_data}
\alias{get_static_mean_data}
\title{Get file of means of static data}
\usage{
get_static_mean_data(old_filename, filename)
}
\arguments{
\item{old_filename}{Name of csv file containing static data}

\item{filename}{Name of new csv file containing means}
}
\value{

}
\description{
Means are taken every second (25 data points).
}
