% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_static.R
\name{hist_plot_static}
\alias{hist_plot_static}
\title{Create histograms from static data}
\usage{
hist_plot_static(data, filename)
}
\arguments{
\item{data}{A dataframe with columns X_static,
Y_static, Z_static}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
Image file containing histograms for X_static, Y_static, Z_static
}
\description{
Create histograms from static data
}
