% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data_static.R
\name{pacf_plot_static}
\alias{pacf_plot_static}
\title{Create PACF plots from static data}
\usage{
pacf_plot_static(data, filename)
}
\arguments{
\item{data}{A dataframe with columns X_static,
Y_static, Z_static}

\item{filename}{String containing the first part of filename for the
image files to be created}
}
\value{
None
}
\description{
Creates an image file containing PACF plots
for X_static, Y_static, Z_static.
}
\examples{
filename <- "Custom_Lady_27Mar17_static.csv"
data <- read.csv(filename)
pacf_plot_static(data, "Lady_27Mar17_pacf")
}
