% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploring_lab_data.R
\name{pairs_plot}
\alias{pairs_plot}
\title{Creates correlation plot}
\usage{
pairs_plot(data, filename)
}
\arguments{
\item{data}{A dataframe}

\item{filename}{Name of image file}
}
\value{
Saves plot as an image file
}
\description{
Creates correlation plot
}
